chatbotJSVercel/requirements.txt
flask==3.0.3
google-generativeai==0.7.2
python-dotenv==1.0.1
